# [SelectionDAG] Disable FastISel for swiftasync functions