@RKSimon 2023-11-01
Do you have any tests that you can add? Particularly the improved Argument lowering handling?