@felipepiovezan 2023-11-02
> Do you have any tests that you can add? Particularly the improved Argument lowering handling?

Yup, good thing you asked, because I've just found a bug with my `OptLevelChanger` version (the "if the optimization level changed" bit is problematic")