@felipepiovezan 2023-11-02
Added a test and tweaked the implementation slightly